#805 A1: split materializer trust-boundary tests
